requirements.txt
langchain-community
langchain
fastapi[all]
llama-index-llms-mistralai
llama-index-embeddings-mistralai
fastapi
uvicorn
pydantic
pyngrok
nest-asyncio
python-dotenv